Backed out of pointer types.
Granted there will be some gnarly technical problems
implementing pointer types (such as *int) in gotables.

The only actual advantage would be availability of nil
(as in nil pointer to indicate e.g. unknown) so as to avoid
a designated out-of-range value (such as -9999 in some
systems) with the risk of it NOT being out-of-range in
one or more domains.

I didn't reject pointer types because they are considered
by Tony Hoare as his Billion Dollar Mistake. I rejected
them because of the complexity of introducing them,
and the availability of an alternative numeric "nil", that
is, NaN.

If you desparately need a nil int, switch to using one of
the float types and use NaN instead.